Fix misc. SonarQube and Coverity findings
Coverity:
- MacInstalledApps: handle return value of FileInputStream.read()
  to avoid potential out-of-range buffer access (CWE-252)
- SystemInfoHelper: null-check getProcess() return before
  dereferencing in printProcesses() (CWE-476)
- PrivilegedUtil: add comment that default FileSystem is a JVM
  singleton that does not need closing (CWE-404 false positive)

SonarQube:
- LinuxOperatingSystem: group regex alternatives with anchors to
  make precedence explicit (S5850)
- PrivilegedUtil: replace volatile Supplier fields with
  AtomicReference (S3077)
- NvmlUtilFFM, NvmlUtilJNA, AdlUtil: replace volatile Set/Map
  fields with AtomicReference (S3077)
- GuidFFM, MacCentralProcessorFFM, LinuxNetworkParamsFFM,
  MacNetworkParamsFFM: cast int operands to long before assignment
  to long parameters (S2184)

+++ oshi-core-java25/src/main/java/oshi/hardware/platform/mac/MacCentralProcessorFFM.java
@@ -147,13 +147,13 @@ final class MacCentralProcessorFFM extends MacCentralProcessor {
                 for (int cpu = 0; cpu < cpuLimit; cpu++) {
                     int offset = cpu * MacSystem.CPU_STATE_MAX;
                     ticks[cpu][TickType.USER.getIndex()] = FormatUtil.getUnsignedInt(
-                            procInfoPtr.getAtIndex(ValueLayout.JAVA_INT, offset + MacSystem.CPU_STATE_USER));
+                            procInfoPtr.getAtIndex(ValueLayout.JAVA_INT, (long) offset + MacSystem.CPU_STATE_USER));
                     ticks[cpu][TickType.NICE.getIndex()] = FormatUtil.getUnsignedInt(
-                            procInfoPtr.getAtIndex(ValueLayout.JAVA_INT, offset + MacSystem.CPU_STATE_NICE));
+                            procInfoPtr.getAtIndex(ValueLayout.JAVA_INT, (long) offset + MacSystem.CPU_STATE_NICE));
                     ticks[cpu][TickType.SYSTEM.getIndex()] = FormatUtil.getUnsignedInt(
-                            procInfoPtr.getAtIndex(ValueLayout.JAVA_INT, offset + MacSystem.CPU_STATE_SYSTEM));
+                            procInfoPtr.getAtIndex(ValueLayout.JAVA_INT, (long) offset + MacSystem.CPU_STATE_SYSTEM));
                     ticks[cpu][TickType.IDLE.getIndex()] = FormatUtil.getUnsignedInt(
-                            procInfoPtr.getAtIndex(ValueLayout.JAVA_INT, offset + MacSystem.CPU_STATE_IDLE));
+                            procInfoPtr.getAtIndex(ValueLayout.JAVA_INT, (long) offset + MacSystem.CPU_STATE_IDLE));
                 }
             } finally {
                 try {
